Do first commit
This firmware is used for testing of Sense Rev4purpose. This
firmware is uploaded in Testing Device, which will be connected
to UART compatible central system. This system will control the
flow of program via UART. Once Testing setup is up and ready,
Sense Rev4 will be used in default mode to test it's operation.

Signed-off-by: Tejas Vasekar <tsvasekar@gmail.com>