Removes an unnecessary variables, add TODOs, add a local enum for state
TODOs added to all the locations where there were errors/optimizations in the code.
A PIR enabled flag wasn't used used, it is removed.
The context passed to the out_gen module is now defined here instead of the
out_gen module and it is passed and retrieved from the out_gen an uint32_t
device_tick_switch_mode isn't called as it is always in the slow mode.
To improve readability, '!' is replaced by 'false ==' in if statements.