Add condition check to see if operation length is greater than ITT
From this commit firmware will check if inter trigger time is less than
total operation time. If it is, PIR will be enabled after completion of
current operation.
For Timer, time_remain is changed from ticks required to 1ms to 1 tick.
As resolution of MS_TIMER is now variable, sometimes it might happen that
it'll take 0 ticks for 1ms. To avoid such a case, time_remain is 1 tick

Signed-off-by: Tejas Vasekar <tsvasekar@gmail.com>